put filters and projects into its own component
